+++ CHANGELOG.md
@@ -21,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed 
 
 - Auto-save reliability
+- Spell rotation combo box sometimes being emptty
 
 ## [4.9.0] - 2023-06-24